pyproject.toml
[tool.poetry]
name = "codereviewai"
version = "0.1.0"
description = ""
authors = ["Artem Svistelnik <artem.svistelnik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
pydantic = "^2.10.4"
pydantic-settings = "^2.7.1"
uvicorn = "^0.34.0"
uvloop = "^0.21.0"
black = "^24.10.0"
redis = "^5.2.1"
github3-py = "^4.0.1"
openai = "^1.59.4"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.2"
httpx = "^0.28.1"
pytest-cov = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
